Move vector and matrix inline functions to where the original headers kept them
The inline bodies of the vector and matrix classes move out of `matrix4d.inl.h` into `matrix.h` and `mxmatrix.h`, the vector length helpers get their own `vectorlength.inl.h`, `roi.h` merges into `realtime.h`, and `realtime.cpp` merges into `orientableroi.cpp`. The original build compiled these inline functions from exactly these places, which is what puts them at the right addresses in the binary. No logic changes.
