+++ LEGO1/realtime/orientableroi.cpp
@@ -1,6 +1,9 @@
+#include "realtime/matrix4d.inl.h"
+#include "realtime/vectorlength.inl.h"
 #include "orientableroi.h"
 
 #include "decomp.h"
+#include "realtime.h"
 
 #include <vec.h>
 
@@ -37,6 +40,7 @@ void OrientableROI::UpdateTransformationRelativeToParent(const Matrix4& p_transf
 	double local2world[4][4];
 	double local2parent[4][4];
 	int i, j;
+	double local_inverse[4][4];
 
 	for (i = 0; i < 4; i++) {
 		for (j = 0; j < 4; j++) {
@@ -45,7 +49,6 @@ void OrientableROI::UpdateTransformationRelativeToParent(const Matrix4& p_transf
 		}
 	}
 
-	double local_inverse[4][4];
 	INVERTMAT4d(local_inverse, local2parent);
 
 	double parent2world[4][4];
@@ -75,6 +78,7 @@ void OrientableROI::GetLocalTransform(Matrix4& p_transform)
 	MxMatrix mat;
 
 	if (m_parentROI != NULL) {
+		double local_inverse[4][4];
 		double local2parent[4][4];
 		unsigned int i, j;
 
@@ -84,7 +88,6 @@ void OrientableROI::GetLocalTransform(Matrix4& p_transform)
 			}
 		}
 
-		double local_inverse[4][4];
 		INVERTMAT4d(local_inverse, local2parent);
 
 		for (i = 0; i < 4; i++) {
@@ -191,6 +194,24 @@ void CalcWorldBoundingVolumes(
 	world_bounding_box.Max()[0] = world_bounding_sphere.Center()[0] + world_bounding_sphere.Radius();
 	world_bounding_box.Max()[1] = world_bounding_sphere.Center()[1] + world_bounding_sphere.Radius();
 	world_bounding_box.Max()[2] = world_bounding_sphere.Center()[2] + world_bounding_sphere.Radius();
+}
+
+// FUNCTION: LEGO1 0x100a5b40
+// FUNCTION: BETA10 0x10168127
+void CalcLocalTransform(const Vector3& p_posVec, const Vector3& p_dirVec, const Vector3& p_upVec, Matrix4& p_outMatrix)
+{
+	float x_axis[3], y_axis[3], z_axis[3];
+
+	NORMVEC3(z_axis, p_dirVec);
+	NORMVEC3(y_axis, p_upVec)
+	VXV3(x_axis, y_axis, z_axis);
+	NORMVEC3(x_axis, x_axis);
+	VXV3(y_axis, z_axis, x_axis);
+	NORMVEC3(y_axis, y_axis);
+	SET4from3(p_outMatrix[0], x_axis, 0);
+	SET4from3(p_outMatrix[1], y_axis, 0);
+	SET4from3(p_outMatrix[2], z_axis, 0);
+	SET4from3(p_outMatrix[3], p_posVec, 1);
 }
 
 // FUNCTION: LEGO1 0x100a5d80
