+++ LEGO1/realtime/vector3d.inl.h
@@ -103,13 +103,6 @@ void Vector3::Clear()
 	memset(m_data, 0, sizeof(float) * 3);
 }
 
-// FUNCTION: LEGO1 0x10003bd0
-// FUNCTION: BETA10 0x10011530
-float Vector3::LenSquared() const
-{
-	return m_data[0] * m_data[0] + m_data[1] * m_data[1] + m_data[2] * m_data[2];
-}
-
 // FUNCTION: LEGO1 0x10003bf0
 // FUNCTION: BETA10 0x100115a0
 void Vector3::Fill(const float& p_value)
