+++ LEGO1/realtime/vector4d.inl.h
@@ -111,40 +111,20 @@ void Vector4::Fill(const float& p_value)
 	m_data[3] = p_value;
 }
 
-// FUNCTION: LEGO1 0x10002b70
-// FUNCTION: BETA10 0x10048ad0
-int Vector4::NormalizeQuaternion()
-{
-	float length = m_data[0] * m_data[0] + m_data[1] * m_data[1] + m_data[2] * m_data[2];
-
-	if (length > 0.0f) {
-		float theta = m_data[3] * 0.5f;
-		float magnitude = sin((double) theta);
-		m_data[3] = cos((double) theta);
-
-		magnitude = magnitude / (float) sqrt((double) length);
-		m_data[0] *= magnitude;
-		m_data[1] *= magnitude;
-		m_data[2] *= magnitude;
-		return 0;
-	}
-	else {
-		return -1;
-	}
-}
+#include "vector.h"
 
 // FUNCTION: LEGO1 0x10002bf0
 // FUNCTION: BETA10 0x10048c20
 int Vector4::EqualsHamiltonProduct(const Vector4& p_a, const Vector4& p_b)
 {
-	m_data[3] = p_a.m_data[3] * p_b.m_data[3] -
-				(p_a.m_data[0] * p_b.m_data[0] + p_a.m_data[2] * p_b.m_data[2] + p_a.m_data[1] * p_b.m_data[1]);
+	m_data[3] = p_b.m_data[3] * p_a.m_data[3] -
+				(p_b.m_data[1] * p_a.m_data[1] + p_b.m_data[2] * p_a.m_data[2] + p_b.m_data[0] * p_a.m_data[0]);
 
 	Vector3::EqualsCrossImpl(p_a.m_data, p_b.m_data);
 
 	m_data[0] = p_b.m_data[3] * p_a.m_data[0] + p_a.m_data[3] * p_b.m_data[0] + m_data[0];
-	m_data[1] = p_b.m_data[1] * p_a.m_data[3] + p_a.m_data[1] * p_b.m_data[3] + m_data[1];
-	m_data[2] = p_b.m_data[2] * p_a.m_data[3] + p_a.m_data[2] * p_b.m_data[3] + m_data[2];
+	m_data[1] = p_a.m_data[1] * p_b.m_data[3] + p_b.m_data[1] * p_a.m_data[3] + m_data[1];
+	m_data[2] = p_a.m_data[2] * p_b.m_data[3] + p_b.m_data[2] * p_a.m_data[3] + m_data[2];
 	return 0;
 }
